Include showTaskList as a module
